minigpt2.0 model修改

+++ utils/bin_dataset.py
@@ -4,7 +4,7 @@ from torch.utils.data import Dataset
 class BinTokenDataset(Dataset):
     def __init__(self, path, seq_len, max_tokens=None):
         import numpy as np
-        raw_data = np.memmap(path, dtype=np.uint16, mode='r')
+        raw_data = np.memmap(path, dtype=np.uint32, mode='r')
         if max_tokens:
             raw_data = raw_data[:max_tokens]
         self.data = raw_data
